requirements.txt
Django==1.5.1
MySQL-python==1.2.4
South==0.7.6
Werkzeug==0.8.3
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-taggit==0.10a1
djangorestframework==2.3.2
ipython==0.13.2
psycopg2==2.5
six==1.3.0
solrpy==0.9.5
wsgiref==0.1.2
